fic merge conflict

+++ api/app/Controllers/Http/IngredientsController.ts
@@ -27,16 +27,13 @@ export default class IngredientsController {
 
     async delete ({params}: HttpContextContract) {
         const ingredient = await Ingredient.findOrFail(params.id)
-
-
-        // someHow this code blocked the DELETE Request 
-        // const recipe = await ingredient.related('recipes').query().first()
-        // if (recipe) {
-        //     throw new Error(`Cet ingrédient est utilisé pour la recette '${recipe.title}'`);
-        // }
-        // await ingredient.related('recipes').detach()
+        const recipe = await ingredient.related('recipes').query().first()
+        if (recipe) {
+            throw new Error(`Cet ingrédient est utilisé pour la recette '${recipe.title}'`);
+        }
+        await ingredient.related('recipes').detach()
         await ingredient.delete()
-        return JSON.stringify({action:true})
+        return null
     }
 
 }
